requirements.txt
django
django-tables2==2.3.1
django-filter==2.2.0
django-crispy-forms==1.9.0
django-autocomplete-light===3.5.1
